API/requirements.txt
scipy==1.5.3
Flask_SQLAlchemy==2.4.4
numpy==1.18.5
python_speech_features==0.6
Flask_RESTful==0.3.8
Flask==1.1.2
Flask_Cors==3.0.9
PyJWT==1.7.1
scikit_learn==0.23.2
